UPDATE: FRONTEND

+++ database/migrations/2025_11_29_223050_create_quizzes_table.php
@@ -21,7 +21,7 @@ return new class extends Migration
             $table->integer('correct_count');
             $table->integer('wrong_count');
             
-            $table->integer('duration');
+            $table->float('duration');
             $table->timestamps();
 
             $table->timestamp('completed_at')->nullable();
